work around issue with null conversation being inserted into conversation adapter
for some reason when a new conversation is created on one device and synced
to another, the synced device will sometimes get a null conversation inserted
into its list instead of an actual conversation. This will make it at least
show up as a blank item in the list and not be clickable so as to avoid a
force close, but does not fix the issue. It is not consistently reproducable
on my devices.