improve message list expansion animation.
Before, we were delaying the alpha animation for the message list so that it didn't start showing up before the expansion animation was done.

This needed to be improved to stop the slight twitch/lag when opening a conversation. Instead of simply delaying the animation, lets just not load the message list until the expansion animation is done. By reducing database overhead during that animation, we can elimate the latency. The user won't notice a difference since the app already delayed showing the conversation.

So difference: delay loading instead of delay showing for decreased latency. Seems to work great.