Cover up the touch sensitivity on the message list.
Basically just quit out of the long click listener if the recycler view is scrolling. I did disable haptic feedback on the long click though, or else this would have made the haptic feedback go off without showing the dialog, which would have been really weird for the user and distracting. I now just force a haptic feedback after the dialog is shown.

The problem was that something was consuming the ACTION_UP event. This could be a major problem for the future, when we need to click individual messages. I would imagine that this is the touchableMovement method, but will have to experiment more with it to find out. Currently, single click listeners do not work.