move to only uploading batches of 1000 messages at a time. 2000 probably would have been alright, but the timeouts would have been on the android side of things, not on the server, so they would be silent.
2500 messages took 7.5 seconds totall to upload for me, with around 6 seconds for the initial upload, that was on very fast wifi too.

So we can guess that batches of 1000 will take about 3 seconds on good wifi, 6 seconds on bad wifi. LTE should be just fine to upload on. 3G might not work great, but i wouldn't have expected it to.

This gives us plenty of buffer room. Our server will handle this change just fine as well, it shouldn't make any difference on that end.