handle firebase reset notification.
FCM can only handle 100 messages in the queue, if it hits that, then all the messages to that device are discarded and never delivered. The next time that device connects, it will be delivered a 'reset' notification.

This commit (flagged off for now) will basically clear the database and reset everything if that happens. There won't be a notification for the user, just want it to happen in the background